requirements.txt
torch==2.0.1
opencv-python==4.5.5.62
speechrecognition==3.8.1
tensorflow==2.10.0
numpy==1.21.5
jellyfish==0.9.0
pillow==9.0.1
transformers==4.31.0
python-dotenv==1.0.0
google-generativeai==0.1.2
tk
